+++ build.gradle
@@ -63,6 +63,10 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-oauth2-client'
     //이메일 라이브러리
     implementation 'org.springframework.boot:spring-boot-starter-mail'
+
+//    엑셀 다운로드 라이브러리
+    implementation 'org.apache.poi:poi:5.2.2'
+    implementation 'org.apache.poi:poi-ooxml:5.2.2'
 }
 
 tasks.named('test') {
